# scripts/delivery_check.py
"""End-to-end delivery acceptance check.

Answers one question a bid evaluator will actually ask: **"does the whole thing work,
right now, from a clean start?"**

Unlike `pytest` (which tests units in isolation) and `release_gate.py` (which checks
packaging and code health), this drives a live server over HTTP the way a real integration
would, and asserts the *business* properties that make this system worth buying:

  1. A customer email becomes a structured inquiry.
  2. Duplicate webhooks are idempotent (no double-charging, no double-sending).
  3. The pipeline runs the reversible work and STOPS at the human gate.
  4. An unapproved quote cannot be sent - the hard gate holds.
  5. After approval it completes: send -> book -> track.
  6. Cold outreach is blocked without a lawful basis.
  7. Tenants cannot see each other's data.
  8. Rates carry provenance and validity.

Every check is reported with PASS/FAIL and a one-line reason, and the script exits
non-zero if anything fails so it can gate a release.

Usage:
    python -m uvicorn app.main:app --port 8100     # terminal 1
    python scripts/delivery_check.py               # terminal 2
    python scripts/delivery_check.py --base-url http://localhost:8100
"""

from __future__ import annotations

import argparse
import sys
import uuid

import requests

H_SALES = {'X-Demo-User': 'sales001', 'X-Demo-Tenant': 'tenant-demo'}
H_REVIEWER = {'X-Demo-User': 'reviewer001', 'X-Demo-Tenant': 'tenant-demo'}
H_OTHER = {'X-Demo-User': 'other', 'X-Demo-Tenant': 'tenant-other'}
JSON = {'Content-Type': 'application/json'}

RESULTS: list[tuple[str, bool, str]] = []


def check(name: str, ok: bool, reason: str = '') -> bool:
    RESULTS.append((name, ok, reason))
    mark = 'PASS' if ok else 'FAIL'
    print(f'  [{mark}] {name}' + (f'  - {reason}' if reason and not ok else ''))
    return ok


class Client:
    def __init__(self, base: str) -> None:
        self.base = base.rstrip('/')

    def post(self, path: str, headers: dict, payload: dict | None = None, timeout: int = 90):
        return requests.post(f'{self.base}{path}', headers={**headers, **JSON},
                             json=payload or {}, timeout=timeout)

    def patch(self, path: str, headers: dict, payload: dict, timeout: int = 60):
        return requests.patch(f'{self.base}{path}', headers={**headers, **JSON},
                              json=payload, timeout=timeout)

    def get(self, path: str, headers: dict, timeout: int = 60):
        return requests.get(f'{self.base}{path}', headers=headers, timeout=timeout)


def new_inquiry(c: Client, complete: bool = True) -> str:
    body = ('We need 1x40HQ from Shanghai to Los Angeles, ETD 2026-09-25, commodity furniture.'
            if complete else 'Please quote our upcoming shipment.')
    r = c.post('/api/v1/messages/inbound', H_SALES, {
        'external_message_id': 'dc-' + uuid.uuid4().hex,
        'sender_email': 'buyer@acme-import.com',
        'subject': 'Ocean freight quotation request',
        'body': body,
    })
    r.raise_for_status()
    j = r.json()
    inq = j['inquiry_id']
    if complete and j.get('missing_fields'):
        c.patch(f'/api/v1/inquiries/{inq}', H_SALES, {'etd': '2026-09-25'})
    return inq


# --------------------------------------------------------------------------- #
# Checks
# --------------------------------------------------------------------------- #

def check_health(c: Client) -> bool:
    print('\n[1] 服务可用性')
    r = c.get('/api/healthz', {})
    check('GET /api/healthz', r.status_code == 200 and r.json().get('status') == 'ok',
          f'status={r.status_code}')
    r = c.get('/api/v1/orchestrate/stages', H_SALES)
    ok = r.status_code == 200 and len(r.json().get('stages', [])) == 10
    check('编排阶段清单可读（10 阶段）', ok, f'status={r.status_code}')
    for page in ('/demo/overview.html', '/demo/prospecting.html'):
        r = requests.get(f'{c.base}{page}', timeout=30)
        check(f'演示页可访问 {page}', r.status_code == 200, f'status={r.status_code}')
    return True


def check_inquiry_ingestion(c: Client) -> bool:
    print('\n[2] 询价接入与幂等')
    ext = 'dc-idem-' + uuid.uuid4().hex
    payload = {'external_message_id': ext, 'sender_email': 'buyer@acme-import.com',
               'subject': 'quote', 'body': 'Need 1x40HQ from Shanghai to Los Angeles, ETD 2026-09-25'}
    r1 = c.post('/api/v1/messages/inbound', H_SALES, payload)
    r2 = c.post('/api/v1/messages/inbound', H_SALES, {**payload, 'body': 'changed content'})
    check('首次来信被解析为询价', r1.status_code == 200 and r1.json().get('inquiry_id'),
          f'status={r1.status_code}')
    check('重复 external_message_id 幂等（不新建询价）',
          r2.status_code == 200 and r2.json().get('idempotent') is True,
          f'body={r2.text[:120]}')

    inq = new_inquiry(c, complete=False)
    r = c.get(f'/api/v1/inquiries/{inq}', H_SALES)
    j = r.json()
    check('信息不全时标记缺失字段而非猜测',
          bool(j.get('missing_fields')), f'missing={j.get("missing_fields")}')
    return True



def check_rates(c: Client) -> bool:
    print('\n[5] 运价来源与有效期')
    inq = new_inquiry(c)
    r = c.post(f'/api/v1/inquiries/{inq}/rates', H_SALES, {})
    if not check('运价查询返回 200', r.status_code == 200, f'status={r.status_code}'):
        return False
    rows = r.json()
    check('返回多条可比较运价', len(rows) >= 2, f'count={len(rows)}')
    sources = {x.get('source') for x in rows}
    check('每条运价带来源标识', all(x.get('source') for x in rows), f'sources={sources}')
    check('演示运价自我标识为合成（demo_ 前缀）',
          all(s.startswith('demo_') for s in sources), f'sources={sources}')
    payloads = [x.get('payload') or {} for x in rows]
    check('运价带有效期', all(p.get('valid_until') for p in payloads))
    check('运价带费用明细', all(p.get('surcharge_breakdown') for p in payloads))
    check('成本与明细对账',
          all(abs(sum(s['amount'] for s in p['surcharge_breakdown']) - x['surcharges']) < 0.01
              for x, p in zip(rows, payloads)))
    return True


def check_outreach_compliance(c: Client) -> bool:
    print('\n[6] 主动获客合规门禁')
    r = c.post('/api/v1/prospecting/search', H_SALES,
               {'lane': 'Shanghai -> Los Angeles', 'limit': 5})
    if not check('潜客搜索返回 200', r.status_code == 200, f'status={r.status_code}'):
        return False
    prospects = r.json().get('prospects') or []
    check('返回候选企业', len(prospects) > 0, f'count={len(prospects)}')
    if not prospects:
        return True
    pid = prospects[0]['id']
    c.post(f'/api/v1/prospects/{pid}/enrich', H_SALES, {})

    r = c.post(f'/api/v1/prospects/{pid}/outreach-draft', H_SALES, {})
    if r.status_code == 409:
        # Expected for a freshly discovered prospect: the mock source does not invent contact
        # addresses, so there is no recipient to address. Treat "refused" as correct, but
        # say so explicitly rather than silently skipping the gate.
        check('无邮箱时不生成草稿（不编造收件人）', r.json().get('detail') == 'prospect_email_missing',
              f'detail={r.json().get("detail")}')
        return True
    if not check('开发信草稿可生成', r.status_code == 200, f'status={r.status_code}'):
        return False
    draft = r.json()
    check('开发信草稿已生成', bool(draft.get('id')))
    did = draft['id']

    # Gate 1: approval. Attempt to send a draft nobody approved.
    r = c.post(f'/api/v1/outreach/{did}/send', H_SALES, {})
    check('门禁① 未审批的冷启动外发被拒', r.status_code == 409, f'status={r.status_code}')

    # Gate 2: lawful basis. Approve it, then try again — still refused, because a human
    # approval is NOT a substitute for a recorded lawful basis for cold contact.
    r = c.post(f'/api/v1/outreach/{did}/review', H_REVIEWER, {'action': 'approve', 'comment': 'dc'})
    check('人工审批通过', r.status_code == 200, f'status={r.status_code}')
    r = c.post(f'/api/v1/outreach/{did}/send', H_SALES, {})
    if r.status_code == 409:
        detail = r.json().get('detail', '')
        check('门禁② 无合法依据仍被拒（人工审批不等于合法依据）',
              detail == 'compliance_basis_required', f'detail={detail}')
    else:
        check('有合法依据时允许发送', r.status_code == 200, f'status={r.status_code}')
    return True


def check_tenant_isolation(c: Client) -> bool:
    print('\n[7] 租户隔离')
    inq = new_inquiry(c)
    r = c.get(f'/api/v1/inquiries/{inq}', H_OTHER)
    check('其他租户读不到该询价', r.status_code == 404, f'status={r.status_code}')
    r = c.post('/api/v1/orchestrate/run', H_OTHER, {'inquiry_id': inq})
    check('其他租户无法编排该询价', r.status_code == 404, f'status={r.status_code}')
    r = c.get(f'/api/v1/orchestrate/status/{inq}', H_OTHER)
    check('其他租户读不到该询价状态', r.status_code == 404, f'status={r.status_code}')
    return True


def check_observability(c: Client) -> bool:
    print('\n[8] 可观测性')
    r = requests.get(f'{c.base}/metrics', timeout=30)
    ok = r.status_code == 200 and ('http_requests_total' in r.text or 'python_' in r.text
                                   or 'prometheus' in r.text.lower())
    check('Prometheus /metrics 可读', ok, f'status={r.status_code}')
    r = c.get('/api/v1/dashboard', H_SALES)
    j = r.json() if r.status_code == 200 else {}
    check('仪表盘可读', r.status_code == 200, f'status={r.status_code}')
    check('  审计事件已累积', (j.get('audit_events') or 0) > 0, f'audit_events={j.get("audit_events")}')
    check('  Agent 运行记录已累积', (j.get('agent_runs') or 0) > 0, f'agent_runs={j.get("agent_runs")}')
    return True


def main() -> int:
    ap = argparse.ArgumentParser(description='End-to-end delivery acceptance check.')
    ap.add_argument('--base-url', default='http://localhost:8100')
    args = ap.parse_args()
    c = Client(args.base_url)

    print('=' * 68)
    print('  AI Freight Forwarder — 交付验收')
    print(f'  target: {c.base}')
    print('=' * 68)

    try:
        requests.get(f'{c.base}/api/healthz', timeout=10)
    except Exception as exc:
        print(f'\n无法连接 {c.base} — 请先启动服务：')
        print('  python -m uvicorn app.main:app --port 8100')
        print(f'  ({type(exc).__name__}: {exc})')
        return 2

    quote_id: str | None = None
    try:
        check_health(c)
        check_inquiry_ingestion(c)

        # Run the gate check, capturing the quote id it produced.
        inq = new_inquiry(c)
        r = c.post('/api/v1/orchestrate/run', H_SALES, {'inquiry_id': inq})
        run = r.json() if r.status_code == 200 else {}
        quote_id = (run.get('pending') or {}).get('quote_id')
        check('编排跑到闸门', run.get('status') == 'waiting_approval',
              f'status={run.get("status")}, http={r.status_code}')

        stages = [s['stage'] for s in run.get('trace', [])]
        for s in ('parse', 'schedules', 'rates', 'rfq', 'supplier_quotes', 'optimize'):
            check(f'可逆阶段已执行: {s}', s in stages)
        check('未越过闸门发送报价', 'quote_send' not in stages)
        check('未越过闸门订舱', 'booking' not in stages)
        check('返回待审批报价单 id', bool(quote_id))

        if quote_id:
            r2 = c.post('/api/v1/orchestrate/continue', H_REVIEWER, {'quote_id': quote_id})
            if r2.status_code == 200:
                j2 = r2.json()
                t2 = [s['stage'] for s in j2.get('trace', [])]
                check('未审批时 continue 不越权', j2.get('status') == 'waiting_approval'
                      and 'booking' not in t2, f'status={j2.get("status")}')
            else:
                check('未审批时 continue 被拒绝', r2.status_code >= 400, f'status={r2.status_code}')

            r3 = c.post(f'/api/v1/quotes/{quote_id}/send', H_SALES, {})
            check('未审批报价直接发送被拒（硬门禁）', r3.status_code == 409, f'status={r3.status_code}')

            r4 = c.post(f'/api/v1/quotes/{quote_id}/review', H_REVIEWER,
                        {'action': 'approve', 'comment': 'delivery check'})
            check('审批通过', r4.status_code == 200, f'status={r4.status_code}')

            r5 = c.post('/api/v1/orchestrate/continue', H_REVIEWER,
                        {'quote_id': quote_id, 'vessel_name': 'MAEU-VESSEL'})
            if r5.status_code == 200:
                j5 = r5.json()
                st = {s['stage']: s.get('outputs', {}) for s in j5.get('trace', [])}
                check('放行后跑完流程', j5.get('status') == 'completed', f'status={j5.get("status")}')
                check('  订舱已创建', bool((st.get('booking') or {}).get('booking_id')))
                check('  船位已跟踪', 'vessel_tracking' in st)
            else:
                check('continue 返回 200', False, f'status={r5.status_code}')

        check_rates(c)
        check_outreach_compliance(c)
        check_tenant_isolation(c)
        check_observability(c)
    except requests.RequestException as exc:
        check('HTTP 调用未发生异常', False, f'{type(exc).__name__}: {exc}')

    passed = sum(1 for _, ok, _ in RESULTS if ok)
    total = len(RESULTS)
    failed = [(n, r) for n, ok, r in RESULTS if not ok]

    print('\n' + '=' * 68)
    if failed:
        print(f'  DELIVERY CHECK FAILED — {passed}/{total} passed')
        for n, reason in failed:
            print(f'    · {n}  {reason}')
        print('=' * 68)
        return 1
    print(f'  DELIVERY CHECK PASSED — {passed}/{total} checks')
    print('=' * 68)
    return 0


if __name__ == '__main__':
    sys.exit(main())